* Reset

+++ Source/Krypton Components/Krypton.Toolkit/Palette Base/PaletteBorder/PaletteBorderInheritForced.cs
@@ -2,18 +2,14 @@
 /*
  * 
  * Original BSD 3-Clause License (https://github.com/ComponentFactory/Krypton/blob/master/LICENSE)
- *  © Component Factory Pty Ltd, 2006 - 2016, All rights reserved.
+ *  © Component Factory Pty Ltd, 2006 - 2016, (Version 4.5.0.0) All rights reserved.
  * 
  *  New BSD 3-Clause License (https://github.com/Krypton-Suite/Standard-Toolkit/blob/master/LICENSE)
  *  Modifications by Peter Wagner(aka Wagnerp) & Simon Coghlan(aka Smurf-IV), et al. 2017 - 2021. All rights reserved. 
  *  
- *  Modified: Monday 12th April, 2021 @ 18:00 GMT
- *
  */
 #endregion
 
-using System.Drawing;
-using System.Diagnostics;
 
 namespace Krypton.Toolkit
 {
@@ -53,7 +49,7 @@ namespace Krypton.Toolkit
         public void SetInherit(IPaletteBorder paletteBorder)
         {
             Debug.Assert(paletteBorder != null);
-            _inherit = paletteBorder; 
+            _inherit = paletteBorder;
         }
         #endregion
 
@@ -188,7 +184,7 @@ namespace Krypton.Toolkit
         /// </summary>
         /// <param name="state">Palette value should be applicable to this state.</param>
         /// <returns>Border rounding.</returns>
-        public override int GetBorderRounding(PaletteState state) => _inherit.GetBorderRounding(state);
+        public override float GetBorderRounding(PaletteState state) => _inherit.GetBorderRounding(state);
 
         /// <summary>
         /// Gets a border image.
