+++ Source/Krypton Components/Krypton.Toolkit/Palette Base/PaletteBorder/PaletteBorderInheritOverride.cs
@@ -2,18 +2,14 @@
 /*
  * 
  * Original BSD 3-Clause License (https://github.com/ComponentFactory/Krypton/blob/master/LICENSE)
- *  © Component Factory Pty Ltd, 2006 - 2016, All rights reserved.
+ *  © Component Factory Pty Ltd, 2006 - 2016, (Version 4.5.0.0) All rights reserved.
  * 
  *  New BSD 3-Clause License (https://github.com/Krypton-Suite/Standard-Toolkit/blob/master/LICENSE)
  *  Modifications by Peter Wagner(aka Wagnerp) & Simon Coghlan(aka Smurf-IV), et al. 2017 - 2021. All rights reserved. 
  *  
- *  Modified: Monday 12th April, 2021 @ 18:00 GMT
- *
  */
 #endregion
 
-using System.Drawing;
-using System.Diagnostics;
 
 namespace Krypton.Toolkit
 {
@@ -65,7 +61,7 @@ namespace Krypton.Toolkit
             _backup = backup;
         }
         #endregion
-        
+
         #region Apply
         /// <summary>
         /// Gets and sets a value indicating if override should be applied.
@@ -81,7 +77,7 @@ namespace Krypton.Toolkit
         public bool Override { get; set; }
 
         #endregion
-        
+
         #region OverrideState
         /// <summary>
         /// Gets and sets the override palette state to use.
@@ -312,11 +308,11 @@ namespace Krypton.Toolkit
         /// </summary>
         /// <param name="state">Palette value should be applicable to this state.</param>
         /// <returns>Border rounding.</returns>
-        public override int GetBorderRounding(PaletteState state)
+        public override float GetBorderRounding(PaletteState state)
         {
             if (Apply)
             {
-                int ret = _primary.GetBorderRounding(Override ? OverrideState : state);
+                float ret = _primary.GetBorderRounding(Override ? OverrideState : state);
 
                 if (ret == -1)
                 {
